Separate common test generation classes/functions
